% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdc.R
\name{cdc_columns}
\alias{cdc_columns}
\title{Get column metadata for a CDC dataset}
\usage{
cdc_columns(dataset_id)
}
\arguments{
\item{dataset_id}{Socrata dataset ID (e.g. "vbim-akqf")}
}
\value{
tibble: field_name, name, datatype, description
}
\description{
Get column metadata for a CDC dataset
}
